Julkaisutaulu (#86)
* Lisätty validointi ja julkaisu

* Lisätty poistetut opintojaksot

* Korjattu testit

* Julkaisuusn parannuksia

+++ eperusteet-ylops-service/src/main/java/fi/vm/sade/eperusteet/ylops/service/lops2019/impl/Lops2019OpintojaksoServiceImpl.java
@@ -2,8 +2,9 @@ package fi.vm.sade.eperusteet.ylops.service.lops2019.impl;
 
 import fi.vm.sade.eperusteet.ylops.domain.KoulutustyyppiToteutus;
 import fi.vm.sade.eperusteet.ylops.domain.lops2019.Lops2019Opintojakso;
-import fi.vm.sade.eperusteet.ylops.domain.lops2019.Lops2019OpintojaksoPoistettu;
+import fi.vm.sade.eperusteet.ylops.domain.lops2019.Lops2019Poistettu;
 import fi.vm.sade.eperusteet.ylops.domain.lops2019.Lops2019OpintojaksonModuuli;
+import fi.vm.sade.eperusteet.ylops.domain.lops2019.PoistetunTyyppi;
 import fi.vm.sade.eperusteet.ylops.domain.ops.Opetussuunnitelma;
 import fi.vm.sade.eperusteet.ylops.dto.KoodiDto;
 import fi.vm.sade.eperusteet.ylops.dto.PoistettuDto;
@@ -191,10 +192,10 @@ public class Lops2019OpintojaksoServiceImpl implements Lops2019OpintojaksoServic
     public void removeOne(Long opsId, Long opintojaksoId) {
         Lops2019Opintojakso opintojakso = getOpintojakso(opsId, opintojaksoId);
         Opetussuunnitelma ops = getOpetussuunnitelma(opsId);
-        Lops2019OpintojaksoPoistettu poistettu = new Lops2019OpintojaksoPoistettu();
+        Lops2019Poistettu poistettu = new Lops2019Poistettu();
         poistettu.setNimi(opintojakso.getNimi());
         poistettu.setOpetussuunnitelma(ops);
-        poistettu.setOpintojakso(opintojaksoId);
+        poistettu.setPoistettu_id(opintojaksoId);
         poistettu.setPalautettu(false);
         opintojaksoPoistettuRepository.save(poistettu);
         ops.getLops2019().getOpintojaksot().remove(opintojakso);
@@ -213,13 +214,17 @@ public class Lops2019OpintojaksoServiceImpl implements Lops2019OpintojaksoServic
 
     @Override
     public Lops2019OpintojaksoDto restore(Long opsId, Long poistettuId) {
-        Lops2019OpintojaksoPoistettu poistettuInfo = opintojaksoPoistettuRepository.findOne(poistettuId);
+        Lops2019Poistettu poistettuInfo = opintojaksoPoistettuRepository.findOne(poistettuId);
         Opetussuunnitelma opetussuunnitelma = getOpetussuunnitelma(opsId);
         if (poistettuInfo.getOpetussuunnitelma() != opetussuunnitelma) {
             throw new BusinessRuleViolationException("vain-oman-voi-palauttaa");
         }
 
-        Lops2019Opintojakso opintojakso = Lops2019Opintojakso.copy(opintojaksoRepository.getLatestNotNull(poistettuInfo.getOpintojakso()));
+        if (!Objects.equals(poistettuInfo.getTyyppi(), PoistetunTyyppi.OPINTOJAKSO)) {
+            throw new BusinessRuleViolationException("voi-palauttaa-vain-opintojaksoja");
+        }
+
+        Lops2019Opintojakso opintojakso = Lops2019Opintojakso.copy(opintojaksoRepository.getLatestNotNull(poistettuInfo.getPoistettu_id()));
         Lops2019OpintojaksoDto result = addOpintojakso(opsId, mapper.map(opintojakso, Lops2019OpintojaksoDto.class));
         opintojaksoPoistettuRepository.delete(poistettuInfo);
         return result;
@@ -228,7 +233,7 @@ public class Lops2019OpintojaksoServiceImpl implements Lops2019OpintojaksoServic
     @Override
     public List<PoistettuDto> getRemoved(Long opsId) {
         Opetussuunnitelma ops = getOpetussuunnitelma(opsId);
-        List<Lops2019OpintojaksoPoistettu> poistetut = opintojaksoPoistettuRepository.findAllByOpetussuunnitelma(ops);
+        List<Lops2019Poistettu> poistetut = opintojaksoPoistettuRepository.findAllByOpetussuunnitelma(ops);
         return mapper.mapAsList(poistetut, PoistettuDto.class);
     }
